+++ api/src/main/java/com/stortor/spring/web/api/core/OrderItemDto.java
@@ -1,11 +1,13 @@
 package com.stortor.spring.web.api.core;
 
+import java.math.BigDecimal;
+
 public class OrderItemDto {
     private Long productId;
     private String productTitle;
     private int quantity;
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 
     public OrderItemDto(ProductDto productDto) {
         this.productId = productDto.getId();
@@ -15,7 +17,7 @@ public class OrderItemDto {
         this.price = productDto.getPrice();
     }
 
-    public OrderItemDto(Long productId, String productTitle, int quantity, int pricePerProduct, int price) {
+    public OrderItemDto(Long productId, String productTitle, int quantity, BigDecimal pricePerProduct, BigDecimal price) {
         this.productId = productId;
         this.productTitle = productTitle;
         this.quantity = quantity;
@@ -50,19 +52,19 @@ public class OrderItemDto {
         this.quantity = quantity;
     }
 
-    public int getPricePerProduct() {
+    public BigDecimal getPricePerProduct() {
         return pricePerProduct;
     }
 
-    public void setPricePerProduct(int pricePerProduct) {
+    public void setPricePerProduct(BigDecimal pricePerProduct) {
         this.pricePerProduct = pricePerProduct;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
