+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -12,7 +12,7 @@ create table products
 (
     id          bigserial primary key,
     title       varchar(255),
-    price       int,
+    price       numeric (8,2),
     category_id bigint not null references categories (id),
     created_at  timestamp default current_timestamp,
     updated_at  timestamp default current_timestamp
@@ -24,26 +24,26 @@ values ('Молочные продукты'),
        ('Овощи');
 
 insert into products (title, price, category_id)
-values ('Milk', 70, 1),
-       ('Water', 30, 1),
-       ('Bread', 40, 1),
-       ('Apples', 100, 1),
-       ('Oranges', 120, 1),
-       ('Lamb', 300, 2),
-       ('Chicken', 200, 3),
-       ('Banana', 150, 1),
-       ('Tomato', 160, 3),
-       ('Chocolate', 50, 2),
-       ('Sugar', 50, 1),
-       ('Salt', 30, 3),
-       ('Cookies', 90, 1),
-       ('Cucumber', 100, 1),
-       ('Coffee', 95, 1),
-       ('Tea', 50, 2),
-       ('Butter', 200, 2),
-       ('Potato', 50, 1),
-       ('Oil', 90, 1),
-       ('Onion', 40, 3);
+values ('Milk', 70.00, 1),
+       ('Water', 30.00, 1),
+       ('Bread', 40.00, 1),
+       ('Apples', 100.00, 1),
+       ('Oranges', 120.00, 1),
+       ('Lamb', 300.00, 2),
+       ('Chicken', 200.00, 3),
+       ('Banana', 150.00, 1),
+       ('Tomato', 160.00, 3),
+       ('Chocolate', 50.00, 2),
+       ('Sugar', 50.00, 1),
+       ('Salt', 30.00, 3),
+       ('Cookies', 90.00, 1),
+       ('Cucumber', 100.00, 1),
+       ('Coffee', 95.00, 1),
+       ('Tea', 50.00, 2),
+       ('Butter', 200.00, 2),
+       ('Potato', 50.00, 1),
+       ('Oil', 90.00, 1),
+       ('Onion', 40.00, 3);
 
 
 drop table orders if exists cascade;
@@ -51,7 +51,7 @@ create table orders
 (
     id          bigserial primary key,
     username    varchar(255) not null,
-    total_price integer      not null,
+    total_price numeric (8,2)      not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -66,11 +66,11 @@ create table order_items
     category_id       bigint references categories (id),
     order_id          bigint references orders (id),
     quantity          integer,
-    price_per_product integer,
-    price             integer,
+    price_per_product numeric (8,2),
+    price             numeric (8,2),
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
 
 insert into orders (username, total_price, address, phone)
-values ('user1', 200, 'address', '46545');
+values ('user1', 200.00, 'address', '46545');
